javascript is toggling diplay
