+++ src/index.html
@@ -7,7 +7,7 @@
     <!-- TODO: enter site and company title -->
     <title>Janky Homepage</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="main.css">
+    <!-- <link rel="stylesheet" href="main.css"> -->
 <style>
 	img {
 		height: 400px;
@@ -25,6 +25,5 @@
         </nav>
     </header>
 <h1>Janky Homepage:</h1>
-<script src="index.js"></script>
 </body>
 </html>
