--- src/js/other.js
@@ -1,1 +1,0 @@
-export const wombat = "wombat";
