+++ src/new-user.html
@@ -7,7 +7,7 @@
     <!-- TODO: enter site and company title -->
     <title>New User</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-<link rel="stylesheet" href="main.css">
+<!-- <link rel="stylesheet" href="main.css"> -->
 </head>
 <style>
 </style>
