+++ src/user-profile.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
 <link rel="stylesheet" href="main.css">
-<script type="javascript" src="./dist/bundle.js"></script>
+<script type="javascript" src="./main.js"></script>
   <title>User Profile</title>
 </head>
 
@@ -28,7 +28,7 @@
     <h3 class="email">
       <a href="dummy@email.com">dummy@email.com</a>
 </h3>
-    <button class="detailsButton" type="button" onclick="toggleUserDetails">Hide Details</button>
+    <button id="detailsButton" class="detailsButton" type="button" onclick="toggleUserDetails(event)">Hide Details</button>
 </hgroup>
 </body>
