+++ src/users-list.html
@@ -8,7 +8,7 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <!-- <script type="javascript" src="./../dist/bundle.js"></script>  -->
 <script type="javascript" src="./dist/bundle.js"></script>
-    <link rel="stylesheet" href="main.css">
+    <link rel="stylesheet" href="./main.css">
 </head>
 
 <body>
@@ -34,7 +34,7 @@
                 <h3 class="email">
                   <a href="dummy@email.com">dummy@email.com</a>
                 </h3>
-                <button class="detailsButton" type="button">Hide Details</button>
+                <button class="detailsButton" type="button" onclick="toggleUserDetails(event)">Hide Details</button>
             </hgroup>
         </li>
         <li>
@@ -45,7 +45,7 @@
                 <h3 class="email">
                     <a href="dummy@email.com">dummy@email.com</a>
                 </h3>
-                <button class="detailsButton" type="button">Hide Details</button>
+                <button class="detailsButton" type="button" onclick="toggleUserDetails(event)">Hide Details</button>
             </hgroup>
         </li>
         <li>
@@ -56,7 +56,7 @@
                 <h3 class="email">
                     <a href="dummy@email.com">dummy@email.com</a>
                 </h3>
-                <button class="detailsButton" type="button">Hide Details</button>
+                <button class="detailsButton" type="button" onclick="toggleUserDetails(event)">Hide Details</button>
             </hgroup>
         </li>
     </ul>
